Features in golem-common
