+++ golem-cli/Cargo.toml
@@ -24,10 +24,10 @@ harness = false
 
 [dependencies]
 golem-client = { path = "../golem-client", version = "0.0.0" }
-golem-common = { path = "../golem-common", version = "0.0.0" }
-golem-rib = { path = "../golem-rib", version = "0.0.0" }
-golem-wasm-ast = { path = "../wasm-ast", version = "0.0.0" }
-golem-wasm-rpc = { path = "../wasm-rpc", version = "0.0.0", default-features = false, features = ["host"] }
+golem-common = { path = "../golem-common", version = "0.0.0", default-features = false }
+golem-rib = { path = "../golem-rib", version = "0.0.0", default-features = false }
+golem-wasm-ast = { path = "../wasm-ast", version = "0.0.0", default-features = false, features = ["analysis"] }
+golem-wasm-rpc = { path = "../wasm-rpc", version = "0.0.0", default-features = false }
 golem-wasm-rpc-stubgen = { path = "../wasm-rpc-stubgen", version = "0.0.0" }
 
 anyhow.workspace = true
