+++ golem-client/Cargo.toml
@@ -13,9 +13,9 @@ include = ["src/**/*", "Cargo.toml", "build.rs", "openapi/**/*"]
 harness = false
 
 [dependencies]
-golem-common = { path = "../golem-common", version = "0.0.0" }
-golem-wasm-ast = { path = "../wasm-ast", version = "0.0.0" }
-golem-wasm-rpc = { path = "../wasm-rpc", version = "0.0.0", default-features = false, features = ["host"] }
+golem-common = { path = "../golem-common", version = "0.0.0", default-features = false }
+golem-wasm-ast = { path = "../wasm-ast", version = "0.0.0", default-features = false, features = ["analysis"] }
+golem-wasm-rpc = { path = "../wasm-rpc", version = "0.0.0", default-features = false }
 
 async-trait = { workspace = true }
 bytes = { workspace = true }
