+++ golem-common/src/client.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::config::RetryConfig;
+use crate::model::RetryConfig;
 use crate::retries::RetryState;
 use dashmap::DashMap;
 use http::Uri;
