+++ golem-common/src/lib.rs
@@ -15,22 +15,45 @@
 use std::fmt;
 use std::fmt::{Display, Formatter};
 
+#[cfg(feature = "tokio")]
 pub mod cache;
+
+#[cfg(feature = "protobuf")]
 pub mod client;
+
+#[cfg(feature = "config")]
 pub mod config;
 
 pub mod golem_version;
+
+#[cfg(feature = "protobuf")]
 pub mod grpc;
+
+#[cfg(feature = "poem")]
 pub mod json_yaml;
+
+#[cfg(feature = "observability")]
 pub mod metrics;
+
 pub mod model;
 pub mod newtype;
+
+#[cfg(feature = "redis")]
 pub mod redis;
+
+#[cfg(feature = "sql")]
 pub mod repo;
+
 pub mod retriable_error;
+
+#[cfg(feature = "tokio")]
 pub mod retries;
+
 pub mod serialization;
+
+#[cfg(feature = "observability")]
 pub mod tracing;
+
 pub mod uri;
 
 #[cfg(test)]
