+++ golem-common/src/model/component_metadata.rs
@@ -24,10 +24,12 @@ use golem_wasm_ast::{
     component::Component,
     IgnoreAllButMetadata,
 };
-use poem_openapi::Object;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Object, Encode, Decode)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct ComponentMetadata {
     pub exports: Vec<AnalysedExport>,
     pub producers: Vec<Producers>,
@@ -42,110 +44,125 @@ impl ComponentMetadata {
 }
 
 #[derive(
-    Debug,
-    Clone,
-    PartialEq,
-    Eq,
-    Hash,
-    Ord,
-    PartialOrd,
-    Serialize,
-    Deserialize,
-    Object,
-    Encode,
-    Decode,
+    Debug, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Serialize, Deserialize, Encode, Decode,
 )]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct ProducerField {
     pub name: String,
     pub values: Vec<VersionedName>,
 }
 
 #[derive(
-    Debug,
-    Clone,
-    PartialEq,
-    Eq,
-    Hash,
-    Ord,
-    PartialOrd,
-    Serialize,
-    Deserialize,
-    Object,
-    Encode,
-    Decode,
+    Debug, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Serialize, Deserialize, Encode, Decode,
 )]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct VersionedName {
     pub name: String,
     pub version: String,
 }
 
-impl From<golem_api_grpc::proto::golem::component::VersionedName> for VersionedName {
-    fn from(value: golem_api_grpc::proto::golem::component::VersionedName) -> Self {
-        Self {
-            name: value.name,
-            version: value.version,
-        }
-    }
-}
-
-impl From<VersionedName> for golem_api_grpc::proto::golem::component::VersionedName {
-    fn from(value: VersionedName) -> Self {
-        Self {
-            name: value.name,
-            version: value.version,
-        }
-    }
-}
-
-impl From<golem_api_grpc::proto::golem::component::ProducerField> for ProducerField {
-    fn from(value: golem_api_grpc::proto::golem::component::ProducerField) -> Self {
-        Self {
-            name: value.name,
-            values: value.values.into_iter().map(|value| value.into()).collect(),
-        }
-    }
-}
-
-impl From<ProducerField> for golem_api_grpc::proto::golem::component::ProducerField {
-    fn from(value: ProducerField) -> Self {
-        Self {
-            name: value.name,
-            values: value.values.into_iter().map(|value| value.into()).collect(),
-        }
-    }
-}
-
 #[derive(
-    Debug,
-    Clone,
-    PartialEq,
-    Eq,
-    Hash,
-    Ord,
-    PartialOrd,
-    Serialize,
-    Deserialize,
-    Object,
-    Encode,
-    Decode,
+    Debug, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Serialize, Deserialize, Encode, Decode,
 )]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct Producers {
     pub fields: Vec<ProducerField>,
 }
 
-impl From<golem_api_grpc::proto::golem::component::Producers> for Producers {
-    fn from(value: golem_api_grpc::proto::golem::component::Producers) -> Self {
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
+pub struct LinearMemory {
+    /// Initial size of the linear memory in bytes
+    pub initial: u64,
+    /// Optional maximal size of the linear memory in bytes
+    pub maximum: Option<u64>,
+}
+
+impl LinearMemory {
+    const PAGE_SIZE: u64 = 65536;
+}
+
+impl From<Mem> for LinearMemory {
+    fn from(value: Mem) -> Self {
         Self {
-            fields: value.fields.into_iter().map(|field| field.into()).collect(),
+            initial: value.mem_type.limits.min * LinearMemory::PAGE_SIZE,
+            maximum: value
+                .mem_type
+                .limits
+                .max
+                .map(|m| m * LinearMemory::PAGE_SIZE),
         }
     }
 }
 
-impl From<Producers> for golem_api_grpc::proto::golem::component::Producers {
-    fn from(value: Producers) -> Self {
-        Self {
-            fields: value.fields.into_iter().map(|field| field.into()).collect(),
+impl From<RawComponentMetadata> for ComponentMetadata {
+    fn from(value: RawComponentMetadata) -> Self {
+        let producers = value
+            .producers
+            .into_iter()
+            .map(|producers| producers.into())
+            .collect::<Vec<_>>();
+
+        let exports = value.exports.into_iter().collect::<Vec<_>>();
+
+        let memories = value.memories.into_iter().map(LinearMemory::from).collect();
+
+        ComponentMetadata {
+            exports,
+            producers,
+            memories,
         }
+    }
+}
+
+// Metadata of Component in terms of golem_wasm_ast types
+pub struct RawComponentMetadata {
+    pub exports: Vec<AnalysedExport>,
+    pub producers: Vec<WasmAstProducers>,
+    pub memories: Vec<Mem>,
+}
+
+impl RawComponentMetadata {
+    pub fn analyse_component(
+        data: &[u8],
+    ) -> Result<RawComponentMetadata, ComponentProcessingError> {
+        let component = Component::<IgnoreAllButMetadata>::from_bytes(data)
+            .map_err(ComponentProcessingError::Parsing)?;
+
+        let producers = component
+            .get_all_producers()
+            .into_iter()
+            .collect::<Vec<_>>();
+
+        let state = AnalysisContext::new(component);
+
+        let mut exports = state
+            .get_top_level_exports()
+            .map_err(ComponentProcessingError::Analysis)?;
+
+        add_resource_drops(&mut exports);
+
+        let exports = exports.into_iter().collect::<Vec<_>>();
+
+        let memories: Vec<Mem> = state
+            .get_all_memories()
+            .map_err(ComponentProcessingError::Analysis)?
+            .into_iter()
+            .collect();
+
+        Ok(RawComponentMetadata {
+            exports,
+            producers,
+            memories,
+        })
     }
 }
 
@@ -202,159 +219,6 @@ impl From<ProducerField> for golem_wasm_ast::metadata::ProducersField {
                 })
                 .collect(),
         }
-    }
-}
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Object, Encode, Decode)]
-pub struct LinearMemory {
-    /// Initial size of the linear memory in bytes
-    pub initial: u64,
-    /// Optional maximal size of the linear memory in bytes
-    pub maximum: Option<u64>,
-}
-
-impl LinearMemory {
-    const PAGE_SIZE: u64 = 65536;
-}
-
-impl From<golem_wasm_ast::core::Mem> for LinearMemory {
-    fn from(value: golem_wasm_ast::core::Mem) -> Self {
-        Self {
-            initial: value.mem_type.limits.min * LinearMemory::PAGE_SIZE,
-            maximum: value
-                .mem_type
-                .limits
-                .max
-                .map(|m| m * LinearMemory::PAGE_SIZE),
-        }
-    }
-}
-
-impl From<golem_api_grpc::proto::golem::component::LinearMemory> for LinearMemory {
-    fn from(value: golem_api_grpc::proto::golem::component::LinearMemory) -> Self {
-        Self {
-            initial: value.initial,
-            maximum: value.maximum,
-        }
-    }
-}
-
-impl From<LinearMemory> for golem_api_grpc::proto::golem::component::LinearMemory {
-    fn from(value: LinearMemory) -> Self {
-        Self {
-            initial: value.initial,
-            maximum: value.maximum,
-        }
-    }
-}
-
-impl From<RawComponentMetadata> for ComponentMetadata {
-    fn from(value: RawComponentMetadata) -> Self {
-        let producers = value
-            .producers
-            .into_iter()
-            .map(|producers| producers.into())
-            .collect::<Vec<_>>();
-
-        let exports = value.exports.into_iter().collect::<Vec<_>>();
-
-        let memories = value.memories.into_iter().map(LinearMemory::from).collect();
-
-        ComponentMetadata {
-            exports,
-            producers,
-            memories,
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::ComponentMetadata> for ComponentMetadata {
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::component::ComponentMetadata,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            exports: value
-                .exports
-                .into_iter()
-                .map(|export| export.try_into())
-                .collect::<Result<_, _>>()?,
-            producers: value
-                .producers
-                .into_iter()
-                .map(|producer| producer.into())
-                .collect(),
-            memories: value
-                .memories
-                .into_iter()
-                .map(|memory| memory.into())
-                .collect(),
-        })
-    }
-}
-
-impl From<ComponentMetadata> for golem_api_grpc::proto::golem::component::ComponentMetadata {
-    fn from(value: ComponentMetadata) -> Self {
-        Self {
-            exports: value
-                .exports
-                .into_iter()
-                .map(|export| export.into())
-                .collect(),
-            producers: value
-                .producers
-                .into_iter()
-                .map(|producer| producer.into())
-                .collect(),
-            memories: value
-                .memories
-                .into_iter()
-                .map(|memory| memory.into())
-                .collect(),
-        }
-    }
-}
-
-// Metadata of Component in terms of golem_wasm_ast types
-pub struct RawComponentMetadata {
-    pub exports: Vec<AnalysedExport>,
-    pub producers: Vec<WasmAstProducers>,
-    pub memories: Vec<Mem>,
-}
-
-impl RawComponentMetadata {
-    pub fn analyse_component(
-        data: &[u8],
-    ) -> Result<RawComponentMetadata, ComponentProcessingError> {
-        let component = Component::<IgnoreAllButMetadata>::from_bytes(data)
-            .map_err(ComponentProcessingError::Parsing)?;
-
-        let producers = component
-            .get_all_producers()
-            .into_iter()
-            .collect::<Vec<_>>();
-
-        let state = AnalysisContext::new(component);
-
-        let mut exports = state
-            .get_top_level_exports()
-            .map_err(ComponentProcessingError::Analysis)?;
-
-        add_resource_drops(&mut exports);
-
-        let exports = exports.into_iter().collect::<Vec<_>>();
-
-        let memories: Vec<Mem> = state
-            .get_all_memories()
-            .map_err(ComponentProcessingError::Analysis)?
-            .into_iter()
-            .collect();
-
-        Ok(RawComponentMetadata {
-            exports,
-            producers,
-            memories,
-        })
     }
 }
 
@@ -427,5 +291,130 @@ fn drop_from_constructor(constructor: &AnalysedFunction) -> AnalysedFunction {
             })
             .collect(),
         results: vec![],
+    }
+}
+
+#[cfg(feature = "protobuf")]
+mod protobuf {
+    use crate::model::component_metadata::{
+        ComponentMetadata, LinearMemory, ProducerField, Producers, VersionedName,
+    };
+
+    impl From<golem_api_grpc::proto::golem::component::VersionedName> for VersionedName {
+        fn from(value: golem_api_grpc::proto::golem::component::VersionedName) -> Self {
+            Self {
+                name: value.name,
+                version: value.version,
+            }
+        }
+    }
+
+    impl From<VersionedName> for golem_api_grpc::proto::golem::component::VersionedName {
+        fn from(value: VersionedName) -> Self {
+            Self {
+                name: value.name,
+                version: value.version,
+            }
+        }
+    }
+
+    impl From<golem_api_grpc::proto::golem::component::ProducerField> for ProducerField {
+        fn from(value: golem_api_grpc::proto::golem::component::ProducerField) -> Self {
+            Self {
+                name: value.name,
+                values: value.values.into_iter().map(|value| value.into()).collect(),
+            }
+        }
+    }
+
+    impl From<ProducerField> for golem_api_grpc::proto::golem::component::ProducerField {
+        fn from(value: ProducerField) -> Self {
+            Self {
+                name: value.name,
+                values: value.values.into_iter().map(|value| value.into()).collect(),
+            }
+        }
+    }
+
+    impl From<golem_api_grpc::proto::golem::component::Producers> for Producers {
+        fn from(value: golem_api_grpc::proto::golem::component::Producers) -> Self {
+            Self {
+                fields: value.fields.into_iter().map(|field| field.into()).collect(),
+            }
+        }
+    }
+
+    impl From<Producers> for golem_api_grpc::proto::golem::component::Producers {
+        fn from(value: Producers) -> Self {
+            Self {
+                fields: value.fields.into_iter().map(|field| field.into()).collect(),
+            }
+        }
+    }
+
+    impl From<golem_api_grpc::proto::golem::component::LinearMemory> for LinearMemory {
+        fn from(value: golem_api_grpc::proto::golem::component::LinearMemory) -> Self {
+            Self {
+                initial: value.initial,
+                maximum: value.maximum,
+            }
+        }
+    }
+
+    impl From<LinearMemory> for golem_api_grpc::proto::golem::component::LinearMemory {
+        fn from(value: LinearMemory) -> Self {
+            Self {
+                initial: value.initial,
+                maximum: value.maximum,
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::ComponentMetadata> for ComponentMetadata {
+        type Error = String;
+
+        fn try_from(
+            value: golem_api_grpc::proto::golem::component::ComponentMetadata,
+        ) -> Result<Self, Self::Error> {
+            Ok(Self {
+                exports: value
+                    .exports
+                    .into_iter()
+                    .map(|export| export.try_into())
+                    .collect::<Result<_, _>>()?,
+                producers: value
+                    .producers
+                    .into_iter()
+                    .map(|producer| producer.into())
+                    .collect(),
+                memories: value
+                    .memories
+                    .into_iter()
+                    .map(|memory| memory.into())
+                    .collect(),
+            })
+        }
+    }
+
+    impl From<ComponentMetadata> for golem_api_grpc::proto::golem::component::ComponentMetadata {
+        fn from(value: ComponentMetadata) -> Self {
+            Self {
+                exports: value
+                    .exports
+                    .into_iter()
+                    .map(|export| export.into())
+                    .collect(),
+                producers: value
+                    .producers
+                    .into_iter()
+                    .map(|producer| producer.into())
+                    .collect(),
+                memories: value
+                    .memories
+                    .into_iter()
+                    .map(|memory| memory.into())
+                    .collect(),
+            }
+        }
     }
 }
