+++ golem-common/src/model/mod.rs
@@ -12,45 +12,34 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::config::RetryConfig;
 use crate::model::oplog::{
     IndexedResourceKey, OplogEntry, OplogIndex, TimestampedUpdateDescription, WorkerResourceId,
 };
 use crate::model::regions::DeletedRegions;
 use crate::newtype_uuid;
 use crate::uri::oss::urn::WorkerUrn;
-use anyhow;
 use bincode::de::read::Reader;
 use bincode::de::{BorrowDecoder, Decoder};
 use bincode::enc::write::Writer;
 use bincode::enc::Encoder;
 use bincode::error::{DecodeError, EncodeError};
 use bincode::{BorrowDecode, Decode, Encode};
-use golem_api_grpc::proto::golem;
-use golem_api_grpc::proto::golem::shardmanager::{
-    Pod as GrpcPod, RoutingTable as GrpcRoutingTable, RoutingTableEntry as GrpcRoutingTableEntry,
-};
-use golem_api_grpc::proto::golem::worker::Cursor;
+
 use golem_wasm_ast::analysis::analysed_type::{
     field, list, r#enum, record, s64, str, tuple, u32, u64,
 };
 use golem_wasm_ast::analysis::{analysed_type, AnalysedType};
 use golem_wasm_rpc::IntoValue;
-use poem::http::Uri;
-use poem_openapi::registry::{MetaSchema, MetaSchemaRef};
-use poem_openapi::types::{ParseFromJSON, ParseFromParameter, ParseResult, ToJSON};
-use poem_openapi::{Enum, NewType, Object, Union};
 use rand::prelude::IteratorRandom;
 use serde::de::Unexpected;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
-use serde_json::Value;
-use std::borrow::Cow;
 use std::cmp::Ordering;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::fmt::{Display, Formatter};
 use std::ops::Add;
 use std::str::FromStr;
 use std::time::{Duration, SystemTime};
+use http::Uri;
 use typed_path::Utf8UnixPathBuf;
 use uuid::{uuid, Uuid};
 
@@ -64,6 +53,30 @@ pub mod plugin;
 pub mod public_oplog;
 pub mod regions;
 pub mod trim_date;
+
+#[cfg(feature = "poem")]
+mod poem;
+
+#[cfg(feature = "protobuf")]
+pub mod protobuf;
+
+#[cfg(feature = "poem")]
+pub trait PoemTypeRequirements:
+    poem_openapi::types::Type + poem_openapi::types::ParseFromJSON + poem_openapi::types::ToJSON
+{
+}
+
+#[cfg(not(feature = "poem"))]
+pub trait PoemTypeRequirements {}
+
+#[cfg(feature = "poem")]
+impl<T: poem_openapi::types::Type + poem_openapi::types::ParseFromJSON + poem_openapi::types::ToJSON>
+    PoemTypeRequirements for T
+{
+}
+
+#[cfg(not(feature = "poem"))]
+impl<T> PoemTypeRequirements for T {}
 
 newtype_uuid!(
     ComponentId,
@@ -96,57 +109,6 @@ impl Timestamp {
 impl Display for Timestamp {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.0)
-    }
-}
-
-impl ToJSON for Timestamp {
-    fn to_json(&self) -> Option<Value> {
-        Some(Value::String(self.0.to_string()))
-    }
-}
-
-impl poem_openapi::types::Type for Timestamp {
-    const IS_REQUIRED: bool = true;
-    type RawValueType = Self;
-    type RawElementValueType = Self;
-
-    fn name() -> Cow<'static, str> {
-        Cow::from("string(timestamp)")
-    }
-
-    fn schema_ref() -> MetaSchemaRef {
-        MetaSchemaRef::Inline(Box::new(MetaSchema::new_with_format("string", "date-time")))
-    }
-
-    fn as_raw_value(&self) -> Option<&Self::RawValueType> {
-        Some(self)
-    }
-
-    fn raw_element_iter<'a>(
-        &'a self,
-    ) -> Box<dyn Iterator<Item = &'a Self::RawElementValueType> + 'a> {
-        Box::new(self.as_raw_value().into_iter())
-    }
-}
-
-impl ParseFromParameter for Timestamp {
-    fn parse_from_parameter(value: &str) -> ParseResult<Self> {
-        value.parse().map_err(|_| {
-            poem_openapi::types::ParseError::<Timestamp>::custom(
-                "Unexpected representation of timestamp".to_string(),
-            )
-        })
-    }
-}
-
-impl ParseFromJSON for Timestamp {
-    fn parse_from_json(value: Option<Value>) -> ParseResult<Self> {
-        match value {
-            Some(Value::String(s)) => Timestamp::parse_from_parameter(&s),
-            _ => Err(poem_openapi::types::ParseError::<Timestamp>::custom(
-                "Unexpected representation of timestamp".to_string(),
-            )),
-        }
     }
 }
 
@@ -216,27 +178,6 @@ impl<'de> bincode::BorrowDecode<'de> for Timestamp {
     }
 }
 
-impl From<Timestamp> for prost_types::Timestamp {
-    fn from(value: Timestamp) -> Self {
-        let d = value
-            .0
-            .duration_since(iso8601_timestamp::Timestamp::UNIX_EPOCH);
-        Self {
-            seconds: d.whole_seconds(),
-            nanos: d.subsec_nanoseconds(),
-        }
-    }
-}
-
-impl From<prost_types::Timestamp> for Timestamp {
-    fn from(value: prost_types::Timestamp) -> Self {
-        Timestamp(
-            iso8601_timestamp::Timestamp::UNIX_EPOCH
-                .add(Duration::new(value.seconds as u64, value.nanos as u32)),
-        )
-    }
-}
-
 impl From<u64> for Timestamp {
     fn from(value: u64) -> Self {
         Timestamp(iso8601_timestamp::Timestamp::UNIX_EPOCH.add(Duration::from_millis(value)))
@@ -261,9 +202,10 @@ impl IntoValue for Timestamp {
 
 pub type ComponentVersion = u64;
 
-#[derive(Clone, Debug, Eq, PartialEq, Hash, Encode, Decode, Serialize, Deserialize, Object)]
+#[derive(Clone, Debug, Eq, PartialEq, Hash, Encode, Decode, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct WorkerId {
     pub component_id: ComponentId,
     pub worker_name: String,
@@ -315,28 +257,6 @@ impl FromStr for WorkerId {
 impl Display for WorkerId {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.write_str(&format!("{}/{}", self.component_id, self.worker_name))
-    }
-}
-
-impl From<WorkerId> for golem_api_grpc::proto::golem::worker::WorkerId {
-    fn from(value: WorkerId) -> Self {
-        Self {
-            component_id: Some(value.component_id.into()),
-            name: value.worker_name,
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::worker::WorkerId> for WorkerId {
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::worker::WorkerId,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            component_id: value.component_id.unwrap().try_into()?,
-            worker_name: value.name,
-        })
     }
 }
 
@@ -481,32 +401,10 @@ impl From<&WorkerId> for TargetWorkerId {
     }
 }
 
-impl TryFrom<golem::worker::TargetWorkerId> for TargetWorkerId {
-    type Error = String;
-
-    fn try_from(value: golem::worker::TargetWorkerId) -> Result<Self, Self::Error> {
-        Ok(Self {
-            component_id: value
-                .component_id
-                .ok_or("Missing component_id")?
-                .try_into()?,
-            worker_name: value.name,
-        })
-    }
-}
-
-impl From<TargetWorkerId> for golem::worker::TargetWorkerId {
-    fn from(value: TargetWorkerId) -> Self {
-        Self {
-            component_id: Some(value.component_id.into()),
-            name: value.worker_name,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Hash, Encode, Decode, Serialize, Deserialize, Object)]
+#[derive(Clone, Debug, Eq, PartialEq, Hash, Encode, Decode, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct PromiseId {
     pub worker_id: WorkerId,
     pub oplog_idx: OplogIndex,
@@ -515,28 +413,6 @@ pub struct PromiseId {
 impl PromiseId {
     pub fn to_redis_key(&self) -> String {
         format!("{}:{}", self.worker_id.to_redis_key(), self.oplog_idx)
-    }
-}
-
-impl From<PromiseId> for golem_api_grpc::proto::golem::worker::PromiseId {
-    fn from(value: PromiseId) -> Self {
-        Self {
-            worker_id: Some(value.worker_id.into()),
-            oplog_idx: value.oplog_idx.into(),
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::worker::PromiseId> for PromiseId {
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::worker::PromiseId,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            worker_id: value.worker_id.ok_or("Missing worker_id")?.try_into()?,
-            oplog_idx: OplogIndex::from_u64(value.oplog_idx),
-        })
     }
 }
 
@@ -622,20 +498,11 @@ impl Display for ScheduleId {
 }
 
 #[derive(
-    Clone,
-    Copy,
-    Debug,
-    Eq,
-    PartialEq,
-    PartialOrd,
-    Ord,
-    Hash,
-    Serialize,
-    Deserialize,
-    Encode,
-    Decode,
-    Object,
+    Clone, Copy, Debug, Eq, PartialEq, PartialOrd, Ord, Hash, Serialize, Deserialize, Encode, Decode,
 )]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct ShardId {
     value: i64,
 }
@@ -684,18 +551,6 @@ impl Display for ShardId {
     }
 }
 
-impl From<ShardId> for golem_api_grpc::proto::golem::shardmanager::ShardId {
-    fn from(value: ShardId) -> golem_api_grpc::proto::golem::shardmanager::ShardId {
-        golem_api_grpc::proto::golem::shardmanager::ShardId { value: value.value }
-    }
-}
-
-impl From<golem_api_grpc::proto::golem::shardmanager::ShardId> for ShardId {
-    fn from(proto: golem_api_grpc::proto::golem::shardmanager::ShardId) -> Self {
-        Self { value: proto.value }
-    }
-}
-
 impl IntoValue for ShardId {
     fn into_value(self) -> golem_wasm_rpc::Value {
         golem_wasm_rpc::Value::S64(self.value)
@@ -728,15 +583,6 @@ impl Pod {
     }
 }
 
-impl From<GrpcPod> for Pod {
-    fn from(value: GrpcPod) -> Self {
-        Self {
-            host: value.host,
-            port: value.port as u16,
-        }
-    }
-}
-
 #[derive(Clone)]
 pub struct RoutingTable {
     pub number_of_shards: NumberOfShards,
@@ -766,34 +612,10 @@ impl RoutingTable {
     }
 }
 
-impl From<GrpcRoutingTable> for RoutingTable {
-    fn from(value: GrpcRoutingTable) -> Self {
-        Self {
-            number_of_shards: NumberOfShards {
-                value: value.number_of_shards as usize,
-            },
-            shard_assignments: value
-                .shard_assignments
-                .into_iter()
-                .map(RoutingTableEntry::from)
-                .map(|routing_table_entry| (routing_table_entry.shard_id, routing_table_entry.pod))
-                .collect(),
-        }
-    }
-}
-
+#[allow(dead_code)]
 pub struct RoutingTableEntry {
     shard_id: ShardId,
     pod: Pod,
-}
-
-impl From<GrpcRoutingTableEntry> for RoutingTableEntry {
-    fn from(value: GrpcRoutingTableEntry) -> Self {
-        Self {
-            shard_id: value.shard_id.unwrap().into(),
-            pod: value.pod.unwrap().into(),
-        }
-    }
 }
 
 #[derive(Clone, Debug, Default)]
@@ -907,18 +729,6 @@ impl<'de> Deserialize<'de> for IdempotencyKey {
     }
 }
 
-impl From<golem_api_grpc::proto::golem::worker::IdempotencyKey> for IdempotencyKey {
-    fn from(proto: golem_api_grpc::proto::golem::worker::IdempotencyKey) -> Self {
-        Self { value: proto.value }
-    }
-}
-
-impl From<IdempotencyKey> for golem_api_grpc::proto::golem::worker::IdempotencyKey {
-    fn from(value: IdempotencyKey) -> Self {
-        Self { value: value.value }
-    }
-}
-
 impl IntoValue for IdempotencyKey {
     fn into_value(self) -> golem_wasm_rpc::Value {
         golem_wasm_rpc::Value::String(self.value)
@@ -926,55 +736,6 @@ impl IntoValue for IdempotencyKey {
 
     fn get_type() -> AnalysedType {
         analysed_type::str()
-    }
-}
-
-impl poem_openapi::types::Type for IdempotencyKey {
-    const IS_REQUIRED: bool = true;
-    type RawValueType = Self;
-    type RawElementValueType = Self;
-
-    fn name() -> Cow<'static, str> {
-        Cow::from(format!("string({})", stringify!(InvocationKey)))
-    }
-
-    fn schema_ref() -> MetaSchemaRef {
-        MetaSchemaRef::Inline(Box::new(MetaSchema::new("string")))
-    }
-
-    fn as_raw_value(&self) -> Option<&Self::RawValueType> {
-        Some(self)
-    }
-
-    fn raw_element_iter<'a>(
-        &'a self,
-    ) -> Box<dyn Iterator<Item = &'a Self::RawElementValueType> + 'a> {
-        Box::new(self.as_raw_value().into_iter())
-    }
-}
-
-impl ParseFromParameter for IdempotencyKey {
-    fn parse_from_parameter(value: &str) -> ParseResult<Self> {
-        Ok(Self {
-            value: value.to_string(),
-        })
-    }
-}
-
-impl ParseFromJSON for IdempotencyKey {
-    fn parse_from_json(value: Option<Value>) -> ParseResult<Self> {
-        match value {
-            Some(Value::String(s)) => Ok(Self { value: s }),
-            _ => Err(poem_openapi::types::ParseError::<IdempotencyKey>::custom(
-                format!("Unexpected representation of {}", stringify!(InvocationKey)),
-            )),
-        }
-    }
-}
-
-impl ToJSON for IdempotencyKey {
-    fn to_json(&self) -> Option<Value> {
-        Some(Value::String(self.value.clone()))
     }
 }
 
@@ -1041,6 +802,17 @@ impl IntoValue for WorkerMetadata {
 pub struct WorkerResourceDescription {
     pub created_at: Timestamp,
     pub indexed_resource_key: Option<IndexedResourceKey>,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize, Encode, Decode)]
+pub struct RetryConfig {
+    pub max_attempts: u32,
+    #[serde(with = "humantime_serde")]
+    pub min_delay: Duration,
+    #[serde(with = "humantime_serde")]
+    pub max_delay: Duration,
+    pub multiplier: f64,
+    pub max_jitter_factor: Option<f64>,
 }
 
 /// Contains status information about a worker according to a given oplog index.
@@ -1188,7 +960,8 @@ pub struct SuccessfulUpdateRecord {
 ///
 /// This is always recorded together with the current oplog index, and it can only be used
 /// as a source of truth if there are no newer oplog entries since the record.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Enum)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 pub enum WorkerStatus {
     /// The worker is running an invoked function
     Running,
@@ -1247,24 +1020,6 @@ impl Display for WorkerStatus {
             WorkerStatus::Retrying => write!(f, "Retrying"),
             WorkerStatus::Failed => write!(f, "Failed"),
             WorkerStatus::Exited => write!(f, "Exited"),
-        }
-    }
-}
-
-impl From<WorkerStatus> for golem_api_grpc::proto::golem::worker::WorkerStatus {
-    fn from(value: WorkerStatus) -> Self {
-        match value {
-            WorkerStatus::Running => golem_api_grpc::proto::golem::worker::WorkerStatus::Running,
-            WorkerStatus::Idle => golem_api_grpc::proto::golem::worker::WorkerStatus::Idle,
-            WorkerStatus::Suspended => {
-                golem_api_grpc::proto::golem::worker::WorkerStatus::Suspended
-            }
-            WorkerStatus::Interrupted => {
-                golem_api_grpc::proto::golem::worker::WorkerStatus::Interrupted
-            }
-            WorkerStatus::Retrying => golem_api_grpc::proto::golem::worker::WorkerStatus::Retrying,
-            WorkerStatus::Failed => golem_api_grpc::proto::golem::worker::WorkerStatus::Failed,
-            WorkerStatus::Exited => golem_api_grpc::proto::golem::worker::WorkerStatus::Exited,
         }
     }
 }
@@ -1405,67 +1160,6 @@ impl From<&str> for AccountId {
     }
 }
 
-impl From<golem_api_grpc::proto::golem::common::AccountId> for AccountId {
-    fn from(proto: golem_api_grpc::proto::golem::common::AccountId) -> Self {
-        Self { value: proto.name }
-    }
-}
-
-impl From<AccountId> for golem_api_grpc::proto::golem::common::AccountId {
-    fn from(value: AccountId) -> Self {
-        golem_api_grpc::proto::golem::common::AccountId { name: value.value }
-    }
-}
-
-impl poem_openapi::types::Type for AccountId {
-    const IS_REQUIRED: bool = true;
-    type RawValueType = Self;
-    type RawElementValueType = Self;
-
-    fn name() -> Cow<'static, str> {
-        Cow::from("string(account_id)")
-    }
-
-    fn schema_ref() -> MetaSchemaRef {
-        MetaSchemaRef::Inline(Box::new(MetaSchema::new("string")))
-    }
-
-    fn as_raw_value(&self) -> Option<&Self::RawValueType> {
-        Some(self)
-    }
-
-    fn raw_element_iter<'a>(
-        &'a self,
-    ) -> Box<dyn Iterator<Item = &'a Self::RawElementValueType> + 'a> {
-        Box::new(self.as_raw_value().into_iter())
-    }
-}
-
-impl ParseFromParameter for AccountId {
-    fn parse_from_parameter(value: &str) -> ParseResult<Self> {
-        Ok(Self {
-            value: value.to_string(),
-        })
-    }
-}
-
-impl ParseFromJSON for AccountId {
-    fn parse_from_json(value: Option<Value>) -> ParseResult<Self> {
-        match value {
-            Some(Value::String(s)) => Ok(Self { value: s }),
-            _ => Err(poem_openapi::types::ParseError::<AccountId>::custom(
-                "Unexpected representation of AccountId".to_string(),
-            )),
-        }
-    }
-}
-
-impl ToJSON for AccountId {
-    fn to_json(&self) -> Option<Value> {
-        Some(Value::String(self.value.clone()))
-    }
-}
-
 impl Display for AccountId {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", &self.value)
@@ -1486,7 +1180,10 @@ pub trait HasAccountId {
     fn account_id(&self) -> AccountId;
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerNameFilter {
     pub comparator: StringFilterComparator,
     pub value: String,
@@ -1504,7 +1201,10 @@ impl Display for WorkerNameFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerStatusFilter {
     pub comparator: FilterComparator,
     pub value: WorkerStatus,
@@ -1522,7 +1222,10 @@ impl Display for WorkerStatusFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerVersionFilter {
     pub comparator: FilterComparator,
     pub value: ComponentVersion,
@@ -1540,7 +1243,10 @@ impl Display for WorkerVersionFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerCreatedAtFilter {
     pub comparator: FilterComparator,
     pub value: Timestamp,
@@ -1558,7 +1264,10 @@ impl Display for WorkerCreatedAtFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerEnvFilter {
     pub name: String,
     pub comparator: StringFilterComparator,
@@ -1581,7 +1290,10 @@ impl Display for WorkerEnvFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerAndFilter {
     pub filters: Vec<WorkerFilter>,
 }
@@ -1606,7 +1318,10 @@ impl Display for WorkerAndFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerOrFilter {
     pub filters: Vec<WorkerFilter>,
 }
@@ -1631,7 +1346,10 @@ impl Display for WorkerOrFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Object)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct WorkerNotFilter {
     filter: Box<WorkerFilter>,
 }
@@ -1650,8 +1368,9 @@ impl Display for WorkerNotFilter {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Union)]
-#[oai(discriminator_name = "type", one_of = true)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Union))]
+#[cfg_attr(feature = "poem", oai(discriminator_name = "type", one_of = true))]
 #[serde(tag = "type")]
 pub enum WorkerFilter {
     Name(WorkerNameFilter),
@@ -1862,148 +1581,8 @@ impl FromStr for WorkerFilter {
     }
 }
 
-impl TryFrom<golem_api_grpc::proto::golem::worker::WorkerFilter> for WorkerFilter {
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::worker::WorkerFilter,
-    ) -> Result<Self, Self::Error> {
-        match value.filter {
-            Some(filter) => match filter {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Name(filter) => Ok(
-                    WorkerFilter::new_name(filter.comparator.try_into()?, filter.value),
-                ),
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Version(filter) => Ok(
-                    WorkerFilter::new_version(filter.comparator.try_into()?, filter.value),
-                ),
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Status(filter) => {
-                    Ok(WorkerFilter::new_status(
-                        filter.comparator.try_into()?,
-                        filter.value.try_into()?,
-                    ))
-                }
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::CreatedAt(filter) => {
-                    let value = filter
-                        .value
-                        .map(|t| t.into())
-                        .ok_or_else(|| "Missing value".to_string())?;
-                    Ok(WorkerFilter::new_created_at(
-                        filter.comparator.try_into()?,
-                        value,
-                    ))
-                }
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Env(filter) => Ok(
-                    WorkerFilter::new_env(filter.name, filter.comparator.try_into()?, filter.value),
-                ),
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Not(filter) => {
-                    let filter = *filter.filter.ok_or_else(|| "Missing filter".to_string())?;
-                    Ok(WorkerFilter::new_not(filter.try_into()?))
-                }
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::And(
-                    golem_api_grpc::proto::golem::worker::WorkerAndFilter { filters },
-                ) => {
-                    let filters = filters.into_iter().map(|f| f.try_into()).collect::<Result<
-                        Vec<WorkerFilter>,
-                        String,
-                    >>(
-                    )?;
-
-                    Ok(WorkerFilter::new_and(filters))
-                }
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Or(
-                    golem_api_grpc::proto::golem::worker::WorkerOrFilter { filters },
-                ) => {
-                    let filters = filters.into_iter().map(|f| f.try_into()).collect::<Result<
-                        Vec<WorkerFilter>,
-                        String,
-                    >>(
-                    )?;
-
-                    Ok(WorkerFilter::new_or(filters))
-                }
-            },
-            None => Err("Missing filter".to_string()),
-        }
-    }
-}
-
-impl From<WorkerFilter> for golem_api_grpc::proto::golem::worker::WorkerFilter {
-    fn from(value: WorkerFilter) -> Self {
-        let filter = match value {
-            WorkerFilter::Name(WorkerNameFilter { comparator, value }) => {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Name(
-                    golem_api_grpc::proto::golem::worker::WorkerNameFilter {
-                        comparator: comparator.into(),
-                        value,
-                    },
-                )
-            }
-            WorkerFilter::Version(WorkerVersionFilter { comparator, value }) => {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Version(
-                    golem_api_grpc::proto::golem::worker::WorkerVersionFilter {
-                        comparator: comparator.into(),
-                        value,
-                    },
-                )
-            }
-            WorkerFilter::Env(WorkerEnvFilter {
-                name,
-                comparator,
-                value,
-            }) => golem_api_grpc::proto::golem::worker::worker_filter::Filter::Env(
-                golem_api_grpc::proto::golem::worker::WorkerEnvFilter {
-                    name,
-                    comparator: comparator.into(),
-                    value,
-                },
-            ),
-            WorkerFilter::Status(WorkerStatusFilter { comparator, value }) => {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Status(
-                    golem_api_grpc::proto::golem::worker::WorkerStatusFilter {
-                        comparator: comparator.into(),
-                        value: value.into(),
-                    },
-                )
-            }
-            WorkerFilter::CreatedAt(WorkerCreatedAtFilter { comparator, value }) => {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::CreatedAt(
-                    golem_api_grpc::proto::golem::worker::WorkerCreatedAtFilter {
-                        value: Some(value.into()),
-                        comparator: comparator.into(),
-                    },
-                )
-            }
-            WorkerFilter::Not(WorkerNotFilter { filter }) => {
-                let f: golem_api_grpc::proto::golem::worker::WorkerFilter = (*filter).into();
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Not(Box::new(
-                    golem_api_grpc::proto::golem::worker::WorkerNotFilter {
-                        filter: Some(Box::new(f)),
-                    },
-                ))
-            }
-            WorkerFilter::And(filter) => {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::And(
-                    golem_api_grpc::proto::golem::worker::WorkerAndFilter {
-                        filters: filter.filters.into_iter().map(|f| f.into()).collect(),
-                    },
-                )
-            }
-            WorkerFilter::Or(filter) => {
-                golem_api_grpc::proto::golem::worker::worker_filter::Filter::Or(
-                    golem_api_grpc::proto::golem::worker::WorkerOrFilter {
-                        filters: filter.filters.into_iter().map(|f| f.into()).collect(),
-                    },
-                )
-            }
-        };
-
-        golem_api_grpc::proto::golem::worker::WorkerFilter {
-            filter: Some(filter),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Enum)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 pub enum StringFilterComparator {
     Equal,
     NotEqual,
@@ -2021,25 +1600,6 @@ impl StringFilterComparator {
             }
             StringFilterComparator::NotLike => {
                 !value1.to_string().contains(value2.to_string().as_str())
-            }
-        }
-    }
-}
-
-impl From<StringFilterComparator> for golem_api_grpc::proto::golem::common::StringFilterComparator {
-    fn from(value: StringFilterComparator) -> Self {
-        match value {
-            StringFilterComparator::Equal => {
-                golem_api_grpc::proto::golem::common::StringFilterComparator::StringEqual
-            }
-            StringFilterComparator::NotEqual => {
-                golem_api_grpc::proto::golem::common::StringFilterComparator::StringNotEqual
-            }
-            StringFilterComparator::Like => {
-                golem_api_grpc::proto::golem::common::StringFilterComparator::StringLike
-            }
-            StringFilterComparator::NotLike => {
-                golem_api_grpc::proto::golem::common::StringFilterComparator::StringNotLike
             }
         }
     }
@@ -2096,7 +1656,8 @@ impl Display for StringFilterComparator {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode, Enum)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Encode, Decode)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 pub enum FilterComparator {
     Equal,
     NotEqual,
@@ -2148,29 +1709,6 @@ impl FromStr for FilterComparator {
     }
 }
 
-impl From<FilterComparator> for golem_api_grpc::proto::golem::common::FilterComparator {
-    fn from(value: FilterComparator) -> Self {
-        match value {
-            FilterComparator::Equal => {
-                golem_api_grpc::proto::golem::common::FilterComparator::Equal
-            }
-            FilterComparator::NotEqual => {
-                golem_api_grpc::proto::golem::common::FilterComparator::NotEqual
-            }
-            FilterComparator::Less => golem_api_grpc::proto::golem::common::FilterComparator::Less,
-            FilterComparator::LessEqual => {
-                golem_api_grpc::proto::golem::common::FilterComparator::LessEqual
-            }
-            FilterComparator::Greater => {
-                golem_api_grpc::proto::golem::common::FilterComparator::Greater
-            }
-            FilterComparator::GreaterEqual => {
-                golem_api_grpc::proto::golem::common::FilterComparator::GreaterEqual
-            }
-        }
-    }
-}
-
 impl TryFrom<i32> for FilterComparator {
     type Error = String;
 
@@ -2200,7 +1738,10 @@ impl From<FilterComparator> for i32 {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Encode, Decode, Object, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Encode, Decode, Default)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct ScanCursor {
     pub cursor: u64,
     pub layer: usize,
@@ -2250,24 +1791,6 @@ impl FromStr for ScanCursor {
     }
 }
 
-impl From<Cursor> for ScanCursor {
-    fn from(value: Cursor) -> Self {
-        Self {
-            cursor: value.cursor,
-            layer: value.layer as usize,
-        }
-    }
-}
-
-impl From<ScanCursor> for Cursor {
-    fn from(value: ScanCursor) -> Self {
-        Self {
-            cursor: value.cursor,
-            layer: value.layer as u64,
-        }
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Encode, Decode, Serialize, Deserialize)]
 pub enum LogLevel {
     Trace,
@@ -2276,32 +1799,6 @@ pub enum LogLevel {
     Warn,
     Error,
     Critical,
-}
-
-impl From<golem_api_grpc::proto::golem::worker::Level> for LogLevel {
-    fn from(value: golem_api_grpc::proto::golem::worker::Level) -> Self {
-        match value {
-            golem_api_grpc::proto::golem::worker::Level::Trace => LogLevel::Trace,
-            golem_api_grpc::proto::golem::worker::Level::Debug => LogLevel::Debug,
-            golem_api_grpc::proto::golem::worker::Level::Info => LogLevel::Info,
-            golem_api_grpc::proto::golem::worker::Level::Warn => LogLevel::Warn,
-            golem_api_grpc::proto::golem::worker::Level::Error => LogLevel::Error,
-            golem_api_grpc::proto::golem::worker::Level::Critical => LogLevel::Critical,
-        }
-    }
-}
-
-impl From<LogLevel> for golem_api_grpc::proto::golem::worker::Level {
-    fn from(value: LogLevel) -> Self {
-        match value {
-            LogLevel::Trace => golem_api_grpc::proto::golem::worker::Level::Trace,
-            LogLevel::Debug => golem_api_grpc::proto::golem::worker::Level::Debug,
-            LogLevel::Info => golem_api_grpc::proto::golem::worker::Level::Info,
-            LogLevel::Warn => golem_api_grpc::proto::golem::worker::Level::Warn,
-            LogLevel::Error => golem_api_grpc::proto::golem::worker::Level::Error,
-            LogLevel::Critical => golem_api_grpc::proto::golem::worker::Level::Critical,
-        }
-    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -2458,135 +1955,8 @@ impl Display for WorkerEvent {
     }
 }
 
-impl TryFrom<golem_api_grpc::proto::golem::worker::LogEvent> for WorkerEvent {
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::worker::LogEvent,
-    ) -> Result<Self, Self::Error> {
-        match value.event {
-            Some(event) => match event {
-                golem_api_grpc::proto::golem::worker::log_event::Event::Stdout(event) => {
-                    Ok(WorkerEvent::StdOut {
-                        timestamp: event.timestamp.ok_or("Missing timestamp")?.into(),
-                        bytes: event.message.into_bytes(),
-                    })
-                }
-                golem_api_grpc::proto::golem::worker::log_event::Event::Stderr(event) => {
-                    Ok(WorkerEvent::StdErr {
-                        timestamp: event.timestamp.ok_or("Missing timestamp")?.into(),
-                        bytes: event.message.into_bytes(),
-                    })
-                }
-                golem_api_grpc::proto::golem::worker::log_event::Event::Log(event) => {
-                    Ok(WorkerEvent::Log {
-                        timestamp: event.timestamp.ok_or("Missing timestamp")?.into(),
-                        level: event.level().into(),
-                        context: event.context,
-                        message: event.message,
-                    })
-                }
-                golem_api_grpc::proto::golem::worker::log_event::Event::InvocationStarted(
-                    event,
-                ) => Ok(WorkerEvent::InvocationStart {
-                    timestamp: event.timestamp.ok_or("Missing timestamp")?.into(),
-                    function: event.function,
-                    idempotency_key: event
-                        .idempotency_key
-                        .ok_or("Missing idempotency key")?
-                        .into(),
-                }),
-                golem_api_grpc::proto::golem::worker::log_event::Event::InvocationFinished(
-                    event,
-                ) => Ok(WorkerEvent::InvocationFinished {
-                    timestamp: event.timestamp.ok_or("Missing timestamp")?.into(),
-                    function: event.function,
-                    idempotency_key: event
-                        .idempotency_key
-                        .ok_or("Missing idempotency key")?
-                        .into(),
-                }),
-            },
-            None => Err("Missing event".to_string()),
-        }
-    }
-}
-
-impl TryFrom<WorkerEvent> for golem_api_grpc::proto::golem::worker::LogEvent {
-    type Error = String;
-
-    fn try_from(value: WorkerEvent) -> Result<Self, Self::Error> {
-        match value {
-            WorkerEvent::StdOut { timestamp, bytes } => Ok(golem::worker::LogEvent {
-                event: Some(golem::worker::log_event::Event::Stdout(
-                    golem::worker::StdOutLog {
-                        message: String::from_utf8_lossy(&bytes).to_string(),
-                        timestamp: Some(timestamp.into()),
-                    },
-                )),
-            }),
-            WorkerEvent::StdErr { timestamp, bytes } => Ok(golem::worker::LogEvent {
-                event: Some(
-                    golem_api_grpc::proto::golem::worker::log_event::Event::Stderr(
-                        golem::worker::StdErrLog {
-                            message: String::from_utf8_lossy(&bytes).to_string(),
-                            timestamp: Some(timestamp.into()),
-                        },
-                    ),
-                ),
-            }),
-            WorkerEvent::Log {
-                timestamp,
-                level,
-                context,
-                message,
-            } => Ok(golem::worker::LogEvent {
-                event: Some(golem::worker::log_event::Event::Log(golem::worker::Log {
-                    level: match level {
-                        LogLevel::Trace => golem::worker::Level::Trace.into(),
-                        LogLevel::Debug => golem::worker::Level::Debug.into(),
-                        LogLevel::Info => golem::worker::Level::Info.into(),
-                        LogLevel::Warn => golem::worker::Level::Warn.into(),
-                        LogLevel::Error => golem::worker::Level::Error.into(),
-                        LogLevel::Critical => golem::worker::Level::Critical.into(),
-                    },
-                    context,
-                    message,
-                    timestamp: Some(timestamp.into()),
-                })),
-            }),
-            WorkerEvent::InvocationStart {
-                timestamp,
-                function,
-                idempotency_key,
-            } => Ok(golem::worker::LogEvent {
-                event: Some(golem::worker::log_event::Event::InvocationStarted(
-                    golem::worker::InvocationStarted {
-                        function,
-                        idempotency_key: Some(idempotency_key.into()),
-                        timestamp: Some(timestamp.into()),
-                    },
-                )),
-            }),
-            WorkerEvent::InvocationFinished {
-                timestamp,
-                function,
-                idempotency_key,
-            } => Ok(golem::worker::LogEvent {
-                event: Some(golem::worker::log_event::Event::InvocationFinished(
-                    golem::worker::InvocationFinished {
-                        function,
-                        idempotency_key: Some(idempotency_key.into()),
-                        timestamp: Some(timestamp.into()),
-                    },
-                )),
-            }),
-            WorkerEvent::Close => Err("Close event is not supported via protobuf".to_string()),
-        }
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, Enum)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 #[repr(i32)]
 pub enum ComponentType {
     Durable = 0,
@@ -2601,32 +1971,6 @@ impl TryFrom<i32> for ComponentType {
             0 => Ok(ComponentType::Durable),
             1 => Ok(ComponentType::Ephemeral),
             _ => Err(format!("Unknown Component Type: {}", value)),
-        }
-    }
-}
-
-impl From<golem_api_grpc::proto::golem::component::ComponentType> for ComponentType {
-    fn from(value: golem_api_grpc::proto::golem::component::ComponentType) -> Self {
-        match value {
-            golem_api_grpc::proto::golem::component::ComponentType::Durable => {
-                ComponentType::Durable
-            }
-            golem_api_grpc::proto::golem::component::ComponentType::Ephemeral => {
-                ComponentType::Ephemeral
-            }
-        }
-    }
-}
-
-impl From<ComponentType> for golem_api_grpc::proto::golem::component::ComponentType {
-    fn from(value: ComponentType) -> Self {
-        match value {
-            ComponentType::Durable => {
-                golem_api_grpc::proto::golem::component::ComponentType::Durable
-            }
-            ComponentType::Ephemeral => {
-                golem_api_grpc::proto::golem::component::ComponentType::Ephemeral
-            }
         }
     }
 }
@@ -2653,12 +1997,16 @@ impl FromStr for ComponentType {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct Empty {}
 
 /// Key that can be used to identify a component file.
 /// All files with the same content will have the same key.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, NewType)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::NewType))]
 pub struct InitialComponentFileKey(pub String);
 
 impl Display for InitialComponentFileKey {
@@ -2716,56 +2064,6 @@ impl Display for ComponentFilePath {
     }
 }
 
-impl poem_openapi::types::Type for ComponentFilePath {
-    const IS_REQUIRED: bool = true;
-
-    type RawValueType = Self;
-
-    type RawElementValueType = Self;
-
-    fn name() -> Cow<'static, str> {
-        "string".into()
-    }
-
-    fn schema_ref() -> MetaSchemaRef {
-        MetaSchemaRef::Inline(Box::new(MetaSchema {
-            description: Some("Path inside a component filesystem. Must be absolute."),
-            ..MetaSchema::new("string")
-        }))
-    }
-
-    fn as_raw_value(&self) -> Option<&Self::RawValueType> {
-        Some(self)
-    }
-
-    fn raw_element_iter<'a>(
-        &'a self,
-    ) -> Box<dyn Iterator<Item = &'a Self::RawElementValueType> + 'a> {
-        Box::new(self.as_raw_value().into_iter())
-    }
-}
-
-impl poem_openapi::types::ToJSON for ComponentFilePath {
-    fn to_json(&self) -> Option<serde_json::Value> {
-        Some(serde_json::Value::String(self.to_string()))
-    }
-}
-
-impl poem_openapi::types::ParseFromJSON for ComponentFilePath {
-    fn parse_from_json(
-        value: Option<serde_json::Value>,
-    ) -> Result<Self, poem_openapi::types::ParseError<Self>> {
-        match value {
-            None => Err(poem_openapi::types::ParseError::custom(
-                "Missing value for ComponentFilePath",
-            )),
-            Some(value) => {
-                serde_json::from_value(value).map_err(poem_openapi::types::ParseError::custom)
-            }
-        }
-    }
-}
-
 impl Serialize for ComponentFilePath {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
@@ -2785,9 +2083,10 @@ impl<'de> Deserialize<'de> for ComponentFilePath {
     }
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize, Enum)]
+#[derive(Clone, Copy, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 #[serde(rename_all = "kebab-case")]
-#[oai(rename_all = "kebab-case")]
+#[cfg_attr(feature = "poem", oai(rename_all = "kebab-case"))]
 pub enum ComponentFilePermissions {
     ReadOnly,
     ReadWrite,
@@ -2809,39 +2108,10 @@ impl ComponentFilePermissions {
     }
 }
 
-impl From<golem_api_grpc::proto::golem::component::ComponentFilePermissions>
-    for ComponentFilePermissions
-{
-    fn from(value: golem_api_grpc::proto::golem::component::ComponentFilePermissions) -> Self {
-        match value {
-            golem_api_grpc::proto::golem::component::ComponentFilePermissions::ReadOnly => {
-                ComponentFilePermissions::ReadOnly
-            }
-            golem_api_grpc::proto::golem::component::ComponentFilePermissions::ReadWrite => {
-                ComponentFilePermissions::ReadWrite
-            }
-        }
-    }
-}
-
-impl From<ComponentFilePermissions>
-    for golem_api_grpc::proto::golem::component::ComponentFilePermissions
-{
-    fn from(value: ComponentFilePermissions) -> Self {
-        match value {
-            ComponentFilePermissions::ReadOnly => {
-                golem_api_grpc::proto::golem::component::ComponentFilePermissions::ReadOnly
-            }
-            ComponentFilePermissions::ReadWrite => {
-                golem_api_grpc::proto::golem::component::ComponentFilePermissions::ReadWrite
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct InitialComponentFile {
     pub key: InitialComponentFileKey,
     pub path: ComponentFilePath,
@@ -2854,42 +2124,10 @@ impl InitialComponentFile {
     }
 }
 
-impl From<InitialComponentFile> for golem_api_grpc::proto::golem::component::InitialComponentFile {
-    fn from(value: InitialComponentFile) -> Self {
-        let permissions: golem_api_grpc::proto::golem::component::ComponentFilePermissions =
-            value.permissions.into();
-        Self {
-            key: value.key.0,
-            path: value.path.to_string(),
-            permissions: permissions.into(),
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::InitialComponentFile>
-    for InitialComponentFile
-{
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::component::InitialComponentFile,
-    ) -> Result<Self, Self::Error> {
-        let permissions: golem_api_grpc::proto::golem::component::ComponentFilePermissions = value
-            .permissions
-            .try_into()
-            .map_err(|e| format!("Failed converting permissions {e}"))?;
-        let permissions: ComponentFilePermissions = permissions.into();
-        let path = ComponentFilePath::from_abs_str(&value.path).map_err(|e| e.to_string())?;
-        let key = InitialComponentFileKey(value.key);
-        Ok(Self {
-            key,
-            path,
-            permissions,
-        })
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, Object)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct ComponentFilePathWithPermissions {
     pub path: ComponentFilePath,
     pub permissions: ComponentFilePermissions,
@@ -2907,7 +2145,10 @@ impl Display for ComponentFilePathWithPermissions {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, Object)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct ComponentFilePathWithPermissionsList {
     pub values: Vec<ComponentFilePathWithPermissions>,
 }
@@ -2915,15 +2156,6 @@ pub struct ComponentFilePathWithPermissionsList {
 impl Display for ComponentFilePathWithPermissionsList {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", serde_json::to_string(self).unwrap())
-    }
-}
-
-impl poem_openapi::types::ParseFromMultipartField for ComponentFilePathWithPermissionsList {
-    async fn parse_from_multipart(field: Option<poem::web::Field>) -> ParseResult<Self> {
-        String::parse_from_multipart(field)
-            .await
-            .map_err(|err| err.propagate::<ComponentFilePathWithPermissionsList>())
-            .and_then(|s| serde_json::from_str(&s).map_err(poem_openapi::types::ParseError::custom))
     }
 }
 
@@ -2943,83 +2175,10 @@ pub struct ComponentFileSystemNode {
     pub details: ComponentFileSystemNodeDetails,
 }
 
-impl From<ComponentFileSystemNode> for golem_api_grpc::proto::golem::worker::FileSystemNode {
-    fn from(value: ComponentFileSystemNode) -> Self {
-        let last_modified = value
-            .last_modified
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
-        match value.details {
-            ComponentFileSystemNodeDetails::File { permissions, size } =>
-                golem_api_grpc::proto::golem::worker::FileSystemNode {
-                    value: Some(golem_api_grpc::proto::golem::worker::file_system_node::Value::File(
-                        golem_api_grpc::proto::golem::worker::FileFileSystemNode {
-                            name: value.name,
-                            last_modified,
-                            size,
-                            permissions:
-                            golem_api_grpc::proto::golem::component::ComponentFilePermissions::from(permissions).into(),
-                        }
-                    ))
-                },
-            ComponentFileSystemNodeDetails::Directory =>
-                golem_api_grpc::proto::golem::worker::FileSystemNode {
-                    value: Some(golem_api_grpc::proto::golem::worker::file_system_node::Value::Directory(
-                        golem_api_grpc::proto::golem::worker::DirectoryFileSystemNode {
-                            name: value.name,
-                            last_modified,
-                        }
-                    ))
-                }
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::worker::FileSystemNode> for ComponentFileSystemNode {
-    type Error = anyhow::Error;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::worker::FileSystemNode,
-    ) -> Result<Self, Self::Error> {
-        match value.value {
-            Some(golem_api_grpc::proto::golem::worker::file_system_node::Value::Directory(
-                golem_api_grpc::proto::golem::worker::DirectoryFileSystemNode {
-                    name,
-                    last_modified,
-                },
-            )) => Ok(ComponentFileSystemNode {
-                name,
-                last_modified: SystemTime::UNIX_EPOCH + Duration::from_secs(last_modified),
-                details: ComponentFileSystemNodeDetails::Directory,
-            }),
-            Some(golem_api_grpc::proto::golem::worker::file_system_node::Value::File(
-                golem_api_grpc::proto::golem::worker::FileFileSystemNode {
-                    name,
-                    last_modified,
-                    size,
-                    permissions,
-                },
-            )) => Ok(ComponentFileSystemNode {
-                name,
-                last_modified: SystemTime::UNIX_EPOCH + Duration::from_secs(last_modified),
-                details: ComponentFileSystemNodeDetails::File {
-                    permissions:
-                        golem_api_grpc::proto::golem::component::ComponentFilePermissions::try_from(
-                            permissions,
-                        )?
-                        .into(),
-                    size,
-                },
-            }),
-            None => Err(anyhow::anyhow!("Missing value")),
-        }
-    }
-}
-#[derive(Debug, Clone, PartialEq, Serialize, Encode, Decode, Enum, Default)]
+#[derive(Debug, Clone, PartialEq, Serialize, Encode, Decode, Default)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 #[serde(rename_all = "kebab-case")]
-#[oai(rename_all = "kebab-case")]
+#[cfg_attr(feature = "poem", oai(rename_all = "kebab-case"))]
 pub enum GatewayBindingType {
     #[default]
     Default,
@@ -3071,41 +2230,6 @@ impl TryFrom<String> for GatewayBindingType {
     }
 }
 
-impl From<golem_api_grpc::proto::golem::apidefinition::GatewayBindingType> for GatewayBindingType {
-    fn from(value: golem_api_grpc::proto::golem::apidefinition::GatewayBindingType) -> Self {
-        match value {
-            golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::Default => {
-                GatewayBindingType::Default
-            }
-            golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::FileServer => {
-                GatewayBindingType::FileServer
-            }
-            golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::CorsPreflight => {
-                GatewayBindingType::CorsPreflight
-            }
-            golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::AuthCallBack => {
-                GatewayBindingType::CorsPreflight
-            }
-        }
-    }
-}
-
-impl From<GatewayBindingType> for golem_api_grpc::proto::golem::apidefinition::GatewayBindingType {
-    fn from(value: GatewayBindingType) -> Self {
-        match value {
-            GatewayBindingType::Default => {
-                golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::Default
-            }
-            GatewayBindingType::FileServer => {
-                golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::FileServer
-            }
-            GatewayBindingType::CorsPreflight => {
-                golem_api_grpc::proto::golem::apidefinition::GatewayBindingType::CorsPreflight
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use test_r::test;
@@ -3125,7 +2249,6 @@ mod tests {
     };
     use bincode::{Decode, Encode};
 
-    use poem_openapi::types::ToJSON;
     use rand::{thread_rng, Rng};
     use serde::{Deserialize, Serialize};
 
@@ -3446,22 +2569,6 @@ mod tests {
     }
 
     #[test]
-    fn worker_status_serialization_poem_serde_equivalence() {
-        let status = WorkerStatus::Retrying;
-        let serialized = status.to_json_string();
-        let deserialized: WorkerStatus = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(status, deserialized);
-    }
-
-    #[test]
-    fn idempotency_key_serialization_poem_serde_equivalence() {
-        let key = IdempotencyKey::fresh();
-        let serialized = key.to_json_string();
-        let deserialized: IdempotencyKey = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(key, deserialized);
-    }
-
-    #[test]
     fn initial_component_file_path_from_absolute() {
         let path = ComponentFilePath::from_abs_str("/a/b/c").unwrap();
         assert_eq!(path.to_string(), "/a/b/c");
@@ -3473,22 +2580,4 @@ mod tests {
         assert!(path.is_err());
     }
 
-    #[test]
-    fn empty_poem_serde_equivalence() {
-        let serialized = Empty {}.to_json_string();
-        let deserialized: Empty = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(Empty {}, deserialized);
-    }
-
-    #[test]
-    fn initial_component_file_serde_equivalence() {
-        let file = InitialComponentFile {
-            key: InitialComponentFileKey("key".to_string()),
-            path: ComponentFilePath::from_rel_str("hello").unwrap(),
-            permissions: ComponentFilePermissions::ReadWrite,
-        };
-        let serialized = file.to_json_string();
-        let deserialized: InitialComponentFile = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(file, deserialized);
-    }
 }
