+++ golem-common/src/model/oplog.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::config::RetryConfig;
+use crate::model::RetryConfig;
 use crate::model::regions::OplogRegion;
 use crate::model::{
     AccountId, ComponentVersion, IdempotencyKey, PluginInstallationId, Timestamp, WorkerId,
@@ -27,7 +27,6 @@ use bincode::{BorrowDecode, Decode, Encode};
 use golem_wasm_ast::analysis::analysed_type::{r#enum, u64};
 use golem_wasm_ast::analysis::AnalysedType;
 use golem_wasm_rpc::{IntoValue, Value};
-use poem_openapi::{Enum, NewType};
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::fmt::{Display, Formatter};
@@ -49,8 +48,8 @@ use uuid::Uuid;
     Encode,
     Decode,
     Default,
-    NewType,
 )]
+#[cfg_attr(feature = "poem", derive(poem_openapi::NewType))]
 pub struct OplogIndex(u64);
 
 impl OplogIndex {
@@ -212,8 +211,8 @@ impl<'de> BorrowDecode<'de> for PayloadId {
     Decode,
     Serialize,
     Deserialize,
-    NewType,
 )]
+#[cfg_attr(feature = "poem", derive(poem_openapi::NewType))]
 pub struct WorkerResourceId(pub u64);
 
 impl WorkerResourceId {
@@ -246,26 +245,9 @@ pub struct IndexedResourceKey {
     pub resource_params: Vec<String>,
 }
 
-impl From<IndexedResourceKey> for golem_api_grpc::proto::golem::worker::IndexedResourceMetadata {
-    fn from(value: IndexedResourceKey) -> Self {
-        golem_api_grpc::proto::golem::worker::IndexedResourceMetadata {
-            resource_name: value.resource_name,
-            resource_params: value.resource_params,
-        }
-    }
-}
-
-impl From<golem_api_grpc::proto::golem::worker::IndexedResourceMetadata> for IndexedResourceKey {
-    fn from(value: golem_api_grpc::proto::golem::worker::IndexedResourceMetadata) -> Self {
-        IndexedResourceKey {
-            resource_name: value.resource_name,
-            resource_params: value.resource_params,
-        }
-    }
-}
-
 /// Worker log levels including the special stdout and stderr channels
-#[derive(Copy, Clone, Debug, PartialEq, Encode, Decode, Serialize, Deserialize, Enum)]
+#[derive(Copy, Clone, Debug, PartialEq, Encode, Decode, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 #[repr(u8)]
 pub enum LogLevel {
     Stdout,
@@ -849,6 +831,29 @@ impl WorkerError {
             WorkerError::InvalidRequest(message) => format!("{message}{error_logs}"),
             WorkerError::StackOverflow => format!("Stack overflow{error_logs}"),
             WorkerError::OutOfMemory => format!("Out of memory{error_logs}"),
+        }
+    }
+}
+
+#[cfg(feature = "protobuf")]
+mod protobuf {
+    use crate::model::oplog::IndexedResourceKey;
+
+    impl From<IndexedResourceKey> for golem_api_grpc::proto::golem::worker::IndexedResourceMetadata {
+        fn from(value: IndexedResourceKey) -> Self {
+            golem_api_grpc::proto::golem::worker::IndexedResourceMetadata {
+                resource_name: value.resource_name,
+                resource_params: value.resource_params,
+            }
+        }
+    }
+
+    impl From<golem_api_grpc::proto::golem::worker::IndexedResourceMetadata> for IndexedResourceKey {
+        fn from(value: golem_api_grpc::proto::golem::worker::IndexedResourceMetadata) -> Self {
+            IndexedResourceKey {
+                resource_name: value.resource_name,
+                resource_params: value.resource_params,
+            }
         }
     }
 }
