+++ golem-common/src/model/plugin.rs
@@ -1,29 +1,24 @@
 use crate::model::{
     AccountId, ComponentId, ComponentVersion, Empty, HasAccountId, PluginInstallationId,
+    PoemTypeRequirements,
 };
-use crate::repo::RowMeta;
 use async_trait::async_trait;
-use poem_openapi::types::{
-    ParseError, ParseFromJSON, ParseFromParameter, ParseResult, ToJSON, Type,
-};
-use poem_openapi::{Enum, Object, Union};
 use serde::{Deserialize, Serialize};
-use sqlx::postgres::PgRow;
-use sqlx::sqlite::SqliteRow;
-use sqlx::{Postgres, Sqlite};
 use std::collections::HashMap;
 use std::fmt::{Debug, Display, Formatter};
 use std::str::FromStr;
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct ComponentPluginScope {
     pub component_id: ComponentId,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Union)]
-#[oai(discriminator_name = "type", one_of = true)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Union))]
+#[cfg_attr(feature = "poem", oai(discriminator_name = "type", one_of = true))]
 #[serde(tag = "type")]
 pub enum DefaultPluginScope {
     Global(Empty),
@@ -62,69 +57,27 @@ impl Display for DefaultPluginScope {
     }
 }
 
-impl ParseFromParameter for DefaultPluginScope {
-    fn parse_from_parameter(value: &str) -> ParseResult<Self> {
+#[cfg(feature = "poem")]
+impl poem_openapi::types::ParseFromParameter for DefaultPluginScope {
+    fn parse_from_parameter(value: &str) -> poem_openapi::types::ParseResult<Self> {
         if value == "global" {
             Ok(Self::global())
         } else if let Some(id_part) = value.strip_prefix("component:") {
             let component_id = ComponentId::try_from(id_part);
             match component_id {
                 Ok(component_id) => Ok(Self::component(component_id)),
-                Err(err) => Err(ParseError::<Self>::custom(err)),
+                Err(err) => Err(poem_openapi::types::ParseError::<Self>::custom(err)),
             }
         } else {
-            Err(ParseError::<Self>::custom("Unexpected representation of plugin scope - must be 'global' or 'component:<component_id>'".to_string()))
+            Err(poem_openapi::types::ParseError::<Self>::custom("Unexpected representation of plugin scope - must be 'global' or 'component:<component_id>'".to_string()))
         }
     }
 }
 
-impl From<DefaultPluginScope> for golem_api_grpc::proto::golem::component::DefaultPluginScope {
-    fn from(scope: DefaultPluginScope) -> Self {
-        match scope {
-            DefaultPluginScope::Global(_) => golem_api_grpc::proto::golem::component::DefaultPluginScope {
-                scope: Some(golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Global(
-                    golem_api_grpc::proto::golem::common::Empty {},
-                )),
-            },
-            DefaultPluginScope::Component(scope) => golem_api_grpc::proto::golem::component::DefaultPluginScope {
-                scope: Some(golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Component(
-                    golem_api_grpc::proto::golem::component::ComponentPluginScope {
-                        component_id: Some(scope.component_id.into()),
-                    },
-                )),
-            },
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::DefaultPluginScope> for DefaultPluginScope {
-    type Error = String;
-
-    fn try_from(
-        proto: golem_api_grpc::proto::golem::component::DefaultPluginScope,
-    ) -> Result<Self, Self::Error> {
-        match proto.scope {
-            Some(golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Global(
-                _,
-            )) => Ok(Self::global()),
-            Some(
-                golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Component(
-                    proto,
-                ),
-            ) => Ok(Self::component(
-                proto
-                    .component_id
-                    .ok_or("Missing component_id".to_string())?
-                    .try_into()?,
-            )),
-            None => Err("Missing scope".to_string()),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct PluginInstallation {
     pub id: PluginInstallationId,
     pub name: String,
@@ -133,37 +86,10 @@ pub struct PluginInstallation {
     pub parameters: HashMap<String, String>,
 }
 
-impl From<PluginInstallation> for golem_api_grpc::proto::golem::component::PluginInstallation {
-    fn from(plugin_installation: PluginInstallation) -> Self {
-        golem_api_grpc::proto::golem::component::PluginInstallation {
-            id: Some(plugin_installation.id.into()),
-            name: plugin_installation.name,
-            version: plugin_installation.version,
-            priority: plugin_installation.priority,
-            parameters: plugin_installation.parameters,
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::PluginInstallation> for PluginInstallation {
-    type Error = String;
-
-    fn try_from(
-        proto: golem_api_grpc::proto::golem::component::PluginInstallation,
-    ) -> Result<Self, Self::Error> {
-        Ok(PluginInstallation {
-            id: proto.id.ok_or("Missing id")?.try_into()?,
-            name: proto.name,
-            version: proto.version,
-            priority: proto.priority,
-            parameters: proto.parameters,
-        })
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct PluginInstallationCreation {
     pub name: String,
     pub version: String,
@@ -183,7 +109,10 @@ impl PluginInstallationCreation {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
+#[serde(rename_all = "camelCase")]
 pub struct PluginInstallationUpdate {
     pub priority: i32,
     pub parameters: HashMap<String, String>,
@@ -196,17 +125,16 @@ pub trait PluginInstallationTarget:
     + PartialEq
     + Serialize
     + for<'de> Deserialize<'de>
-    + Type
-    + ParseFromJSON
-    + ToJSON
+    + PoemTypeRequirements
     + Send
     + Sync
     + 'static
 {
-    type Row: RowMeta<sqlx::Sqlite>
-        + RowMeta<sqlx::Postgres>
-        + for<'r> sqlx::FromRow<'r, SqliteRow>
-        + for<'r> sqlx::FromRow<'r, PgRow>
+    #[cfg(feature = "sql")]
+    type Row: crate::repo::RowMeta<sqlx::Sqlite>
+        + crate::repo::RowMeta<sqlx::Postgres>
+        + for<'r> sqlx::FromRow<'r, sqlx::sqlite::SqliteRow>
+        + for<'r> sqlx::FromRow<'r, sqlx::postgres::PgRow>
         + From<Self>
         + TryInto<Self, Error = String>
         + Clone
@@ -216,6 +144,7 @@ pub trait PluginInstallationTarget:
         + Unpin
         + 'static;
 
+    #[cfg(feature = "sql")]
     fn table_name() -> &'static str;
 }
 
@@ -228,17 +157,16 @@ pub trait PluginOwner:
     + PartialEq
     + Serialize
     + for<'de> Deserialize<'de>
-    + Type
-    + ParseFromJSON
-    + ToJSON
+    + PoemTypeRequirements
     + Send
     + Sync
     + 'static
 {
-    type Row: RowMeta<sqlx::Sqlite>
-        + RowMeta<sqlx::Postgres>
-        + for<'r> sqlx::FromRow<'r, SqliteRow>
-        + for<'r> sqlx::FromRow<'r, PgRow>
+    #[cfg(feature = "sql")]
+    type Row: crate::repo::RowMeta<sqlx::Sqlite>
+        + crate::repo::RowMeta<sqlx::Postgres>
+        + for<'r> sqlx::FromRow<'r, sqlx::sqlite::SqliteRow>
+        + for<'r> sqlx::FromRow<'r, sqlx::postgres::PgRow>
         + From<Self>
         + TryInto<Self, Error = String>
         + Clone
@@ -250,9 +178,10 @@ pub trait PluginOwner:
         + 'static;
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct DefaultPluginOwner;
 
 impl Display for DefaultPluginOwner {
@@ -280,12 +209,14 @@ impl HasAccountId for DefaultPluginOwner {
 }
 
 impl PluginOwner for DefaultPluginOwner {
+    #[cfg(feature = "sql")]
     type Row = crate::repo::plugin::DefaultPluginOwnerRow;
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct PluginDefinition<Owner: PluginOwner, Scope: PluginScope> {
     pub name: String,
     pub version: String,
@@ -297,46 +228,10 @@ pub struct PluginDefinition<Owner: PluginOwner, Scope: PluginScope> {
     pub owner: Owner,
 }
 
-impl From<PluginDefinition<DefaultPluginOwner, DefaultPluginScope>>
-    for golem_api_grpc::proto::golem::component::PluginDefinition
-{
-    fn from(value: PluginDefinition<DefaultPluginOwner, DefaultPluginScope>) -> Self {
-        golem_api_grpc::proto::golem::component::PluginDefinition {
-            name: value.name,
-            version: value.version,
-            description: value.description,
-            icon: value.icon,
-            homepage: value.homepage,
-            specs: Some(value.specs.into()),
-            scope: Some(value.scope.into()),
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::PluginDefinition>
-    for PluginDefinition<DefaultPluginOwner, DefaultPluginScope>
-{
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::component::PluginDefinition,
-    ) -> Result<Self, Self::Error> {
-        Ok(PluginDefinition {
-            name: value.name,
-            version: value.version,
-            description: value.description,
-            icon: value.icon,
-            homepage: value.homepage,
-            specs: value.specs.ok_or("Missing plugin specs")?.try_into()?,
-            scope: value.scope.ok_or("Missing plugin scope")?.try_into()?,
-            owner: DefaultPluginOwner,
-        })
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct PluginDefinitionWithoutOwner<Scope: PluginScope> {
     pub name: String,
     pub version: String,
@@ -362,15 +257,17 @@ impl<Scope: PluginScope> PluginDefinitionWithoutOwner<Scope> {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Enum)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Enum))]
 #[repr(i8)]
 pub enum PluginType {
     ComponentTransformer = 0,
     OplogProcessor = 1,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Union)]
-#[oai(discriminator_name = "type", one_of = true)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Union))]
+#[cfg_attr(feature = "poem", oai(discriminator_name = "type", one_of = true))]
 #[serde(tag = "type")]
 pub enum PluginTypeSpecificDefinition {
     ComponentTransformer(ComponentTransformerDefinition),
@@ -388,39 +285,10 @@ impl PluginTypeSpecificDefinition {
     }
 }
 
-impl From<PluginTypeSpecificDefinition>
-    for golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition
-{
-    fn from(value: PluginTypeSpecificDefinition) -> Self {
-        match value {
-            PluginTypeSpecificDefinition::ComponentTransformer(value) => golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition {
-                definition: Some(golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::ComponentTransformer(value.into()))
-            },
-            PluginTypeSpecificDefinition::OplogProcessor(value) => golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition {
-                definition: Some(golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::OplogProcessor(value.into()))
-            }
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition>
-    for PluginTypeSpecificDefinition
-{
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition,
-    ) -> Result<Self, Self::Error> {
-        match value.definition.ok_or("Missing plugin type specific definition")? {
-            golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::ComponentTransformer(value) => Ok(PluginTypeSpecificDefinition::ComponentTransformer(value.try_into()?)),
-            golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::OplogProcessor(value) => Ok(PluginTypeSpecificDefinition::OplogProcessor(value.try_into()?)),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct ComponentTransformerDefinition {
     pub provided_wit_package: Option<String>,
     pub json_schema: Option<String>,
@@ -428,71 +296,13 @@ pub struct ComponentTransformerDefinition {
     pub transform_url: String,
 }
 
-impl From<ComponentTransformerDefinition>
-    for golem_api_grpc::proto::golem::component::ComponentTransformerDefinition
-{
-    fn from(value: ComponentTransformerDefinition) -> Self {
-        golem_api_grpc::proto::golem::component::ComponentTransformerDefinition {
-            provided_wit_package: value.provided_wit_package,
-            json_schema: value.json_schema,
-            validate_url: value.validate_url,
-            transform_url: value.transform_url,
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::ComponentTransformerDefinition>
-    for ComponentTransformerDefinition
-{
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::component::ComponentTransformerDefinition,
-    ) -> Result<Self, Self::Error> {
-        Ok(ComponentTransformerDefinition {
-            provided_wit_package: value.provided_wit_package,
-            json_schema: value.json_schema,
-            validate_url: value.validate_url,
-            transform_url: value.transform_url,
-        })
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct OplogProcessorDefinition {
     pub component_id: ComponentId,
     pub component_version: ComponentVersion,
-}
-
-impl From<OplogProcessorDefinition>
-    for golem_api_grpc::proto::golem::component::OplogProcessorDefinition
-{
-    fn from(value: OplogProcessorDefinition) -> Self {
-        golem_api_grpc::proto::golem::component::OplogProcessorDefinition {
-            component_id: Some(value.component_id.into()),
-            component_version: value.component_version,
-        }
-    }
-}
-
-impl TryFrom<golem_api_grpc::proto::golem::component::OplogProcessorDefinition>
-    for OplogProcessorDefinition
-{
-    type Error = String;
-
-    fn try_from(
-        value: golem_api_grpc::proto::golem::component::OplogProcessorDefinition,
-    ) -> Result<Self, Self::Error> {
-        Ok(OplogProcessorDefinition {
-            component_id: value
-                .component_id
-                .ok_or("Missing component_id")?
-                .try_into()?,
-            component_version: value.component_version,
-        })
-    }
 }
 
 #[async_trait]
@@ -502,17 +312,16 @@ pub trait PluginScope:
     + PartialEq
     + Serialize
     + for<'de> Deserialize<'de>
-    + Type
-    + ParseFromJSON
-    + ToJSON
+    + PoemTypeRequirements
     + Send
     + Sync
     + 'static
 {
-    type Row: RowMeta<Sqlite>
-        + RowMeta<Postgres>
-        + for<'r> sqlx::FromRow<'r, SqliteRow>
-        + for<'r> sqlx::FromRow<'r, PgRow>
+    #[cfg(feature = "sql")]
+    type Row: crate::repo::RowMeta<sqlx::Sqlite>
+        + crate::repo::RowMeta<sqlx::Postgres>
+        + for<'r> sqlx::FromRow<'r, sqlx::sqlite::SqliteRow>
+        + for<'r> sqlx::FromRow<'r, sqlx::postgres::PgRow>
         + From<Self>
         + TryInto<Self, Error = String>
         + Send
@@ -529,6 +338,7 @@ pub trait PluginScope:
 
 #[async_trait]
 impl PluginScope for DefaultPluginScope {
+    #[cfg(feature = "sql")]
     type Row = crate::repo::plugin::DefaultPluginScopeRow;
 
     type RequestContext = ();
@@ -541,9 +351,10 @@ impl PluginScope for DefaultPluginScope {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Object)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(feature = "poem", derive(poem_openapi::Object))]
+#[cfg_attr(feature = "poem", oai(rename_all = "camelCase"))]
 #[serde(rename_all = "camelCase")]
-#[oai(rename_all = "camelCase")]
 pub struct ComponentPluginInstallationTarget {
     pub component_id: ComponentId,
     pub component_version: ComponentVersion,
@@ -556,9 +367,218 @@ impl Display for ComponentPluginInstallationTarget {
 }
 
 impl PluginInstallationTarget for ComponentPluginInstallationTarget {
+    #[cfg(feature = "sql")]
     type Row = crate::repo::plugin_installation::ComponentPluginInstallationRow;
 
+    #[cfg(feature = "sql")]
     fn table_name() -> &'static str {
         "component_plugin_installation"
+    }
+}
+
+#[cfg(feature = "protobuf")]
+mod protobuf {
+    use crate::model::plugin::{
+        ComponentTransformerDefinition, DefaultPluginOwner, DefaultPluginScope,
+        OplogProcessorDefinition, PluginDefinition, PluginInstallation,
+        PluginTypeSpecificDefinition,
+    };
+
+    impl From<DefaultPluginScope> for golem_api_grpc::proto::golem::component::DefaultPluginScope {
+        fn from(scope: DefaultPluginScope) -> Self {
+            match scope {
+                DefaultPluginScope::Global(_) => golem_api_grpc::proto::golem::component::DefaultPluginScope {
+                    scope: Some(golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Global(
+                        golem_api_grpc::proto::golem::common::Empty {},
+                    )),
+                },
+                DefaultPluginScope::Component(scope) => golem_api_grpc::proto::golem::component::DefaultPluginScope {
+                    scope: Some(golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Component(
+                        golem_api_grpc::proto::golem::component::ComponentPluginScope {
+                            component_id: Some(scope.component_id.into()),
+                        },
+                    )),
+                },
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::DefaultPluginScope> for DefaultPluginScope {
+        type Error = String;
+
+        fn try_from(
+            proto: golem_api_grpc::proto::golem::component::DefaultPluginScope,
+        ) -> Result<Self, Self::Error> {
+            match proto.scope {
+                Some(
+                    golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Global(_),
+                ) => Ok(Self::global()),
+                Some(
+                    golem_api_grpc::proto::golem::component::default_plugin_scope::Scope::Component(
+                        proto,
+                    ),
+                ) => Ok(Self::component(
+                    proto
+                        .component_id
+                        .ok_or("Missing component_id".to_string())?
+                        .try_into()?,
+                )),
+                None => Err("Missing scope".to_string()),
+            }
+        }
+    }
+
+    impl From<PluginDefinition<DefaultPluginOwner, DefaultPluginScope>>
+        for golem_api_grpc::proto::golem::component::PluginDefinition
+    {
+        fn from(value: PluginDefinition<DefaultPluginOwner, DefaultPluginScope>) -> Self {
+            golem_api_grpc::proto::golem::component::PluginDefinition {
+                name: value.name,
+                version: value.version,
+                description: value.description,
+                icon: value.icon,
+                homepage: value.homepage,
+                specs: Some(value.specs.into()),
+                scope: Some(value.scope.into()),
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::PluginDefinition>
+        for PluginDefinition<DefaultPluginOwner, DefaultPluginScope>
+    {
+        type Error = String;
+
+        fn try_from(
+            value: golem_api_grpc::proto::golem::component::PluginDefinition,
+        ) -> Result<Self, Self::Error> {
+            Ok(PluginDefinition {
+                name: value.name,
+                version: value.version,
+                description: value.description,
+                icon: value.icon,
+                homepage: value.homepage,
+                specs: value.specs.ok_or("Missing plugin specs")?.try_into()?,
+                scope: value.scope.ok_or("Missing plugin scope")?.try_into()?,
+                owner: DefaultPluginOwner,
+            })
+        }
+    }
+
+    impl From<PluginInstallation> for golem_api_grpc::proto::golem::component::PluginInstallation {
+        fn from(plugin_installation: PluginInstallation) -> Self {
+            golem_api_grpc::proto::golem::component::PluginInstallation {
+                id: Some(plugin_installation.id.into()),
+                name: plugin_installation.name,
+                version: plugin_installation.version,
+                priority: plugin_installation.priority,
+                parameters: plugin_installation.parameters,
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::PluginInstallation> for PluginInstallation {
+        type Error = String;
+
+        fn try_from(
+            proto: golem_api_grpc::proto::golem::component::PluginInstallation,
+        ) -> Result<Self, Self::Error> {
+            Ok(PluginInstallation {
+                id: proto.id.ok_or("Missing id")?.try_into()?,
+                name: proto.name,
+                version: proto.version,
+                priority: proto.priority,
+                parameters: proto.parameters,
+            })
+        }
+    }
+
+    impl From<PluginTypeSpecificDefinition>
+        for golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition
+    {
+        fn from(value: PluginTypeSpecificDefinition) -> Self {
+            match value {
+                PluginTypeSpecificDefinition::ComponentTransformer(value) => golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition {
+                    definition: Some(golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::ComponentTransformer(value.into()))
+                },
+                PluginTypeSpecificDefinition::OplogProcessor(value) => golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition {
+                    definition: Some(golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::OplogProcessor(value.into()))
+                }
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition>
+        for PluginTypeSpecificDefinition
+    {
+        type Error = String;
+
+        fn try_from(
+            value: golem_api_grpc::proto::golem::component::PluginTypeSpecificDefinition,
+        ) -> Result<Self, Self::Error> {
+            match value.definition.ok_or("Missing plugin type specific definition")? {
+                golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::ComponentTransformer(value) => Ok(PluginTypeSpecificDefinition::ComponentTransformer(value.try_into()?)),
+                golem_api_grpc::proto::golem::component::plugin_type_specific_definition::Definition::OplogProcessor(value) => Ok(PluginTypeSpecificDefinition::OplogProcessor(value.try_into()?)),
+            }
+        }
+    }
+
+    impl From<ComponentTransformerDefinition>
+        for golem_api_grpc::proto::golem::component::ComponentTransformerDefinition
+    {
+        fn from(value: ComponentTransformerDefinition) -> Self {
+            golem_api_grpc::proto::golem::component::ComponentTransformerDefinition {
+                provided_wit_package: value.provided_wit_package,
+                json_schema: value.json_schema,
+                validate_url: value.validate_url,
+                transform_url: value.transform_url,
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::ComponentTransformerDefinition>
+        for ComponentTransformerDefinition
+    {
+        type Error = String;
+
+        fn try_from(
+            value: golem_api_grpc::proto::golem::component::ComponentTransformerDefinition,
+        ) -> Result<Self, Self::Error> {
+            Ok(ComponentTransformerDefinition {
+                provided_wit_package: value.provided_wit_package,
+                json_schema: value.json_schema,
+                validate_url: value.validate_url,
+                transform_url: value.transform_url,
+            })
+        }
+    }
+
+    impl From<OplogProcessorDefinition>
+        for golem_api_grpc::proto::golem::component::OplogProcessorDefinition
+    {
+        fn from(value: OplogProcessorDefinition) -> Self {
+            golem_api_grpc::proto::golem::component::OplogProcessorDefinition {
+                component_id: Some(value.component_id.into()),
+                component_version: value.component_version,
+            }
+        }
+    }
+
+    impl TryFrom<golem_api_grpc::proto::golem::component::OplogProcessorDefinition>
+        for OplogProcessorDefinition
+    {
+        type Error = String;
+
+        fn try_from(
+            value: golem_api_grpc::proto::golem::component::OplogProcessorDefinition,
+        ) -> Result<Self, Self::Error> {
+            Ok(OplogProcessorDefinition {
+                component_id: value
+                    .component_id
+                    .ok_or("Missing component_id")?
+                    .try_into()?,
+                component_version: value.component_version,
+            })
+        }
     }
 }
