+++ golem-common/src/newtype.rs
@@ -68,6 +68,7 @@ macro_rules! newtype_uuid {
             }
         }
 
+        #[cfg(feature = "protobuf")]
         impl TryFrom<$proto_type> for $name {
             type Error = String;
 
@@ -81,6 +82,7 @@ macro_rules! newtype_uuid {
             }
         }
 
+        #[cfg(feature = "protobuf")]
         impl From<$name> for $proto_type {
             fn from(value: $name) -> Self {
                 $proto_type {
@@ -89,17 +91,18 @@ macro_rules! newtype_uuid {
             }
         }
 
+        #[cfg(feature = "poem")]
         impl poem_openapi::types::Type for $name {
             const IS_REQUIRED: bool = true;
             type RawValueType = Self;
             type RawElementValueType = Self;
 
-            fn name() -> Cow<'static, str> {
-                Cow::from(format!("string({})", stringify!($name)))
+            fn name() -> std::borrow::Cow<'static, str> {
+                std::borrow::Cow::from(format!("string({})", stringify!($name)))
             }
 
-            fn schema_ref() -> MetaSchemaRef {
-                MetaSchemaRef::Inline(Box::new(MetaSchema::new_with_format("string", "uuid")))
+            fn schema_ref() -> poem_openapi::registry::MetaSchemaRef {
+                poem_openapi::registry::MetaSchemaRef::Inline(Box::new(poem_openapi::registry::MetaSchema::new_with_format("string", "uuid")))
             }
 
             fn as_raw_value(&self) -> Option<&Self::RawValueType> {
@@ -113,16 +116,18 @@ macro_rules! newtype_uuid {
             }
         }
 
-        impl ParseFromParameter for $name {
-            fn parse_from_parameter(value: &str) -> ParseResult<Self> {
+        #[cfg(feature = "poem")]
+        impl poem_openapi::types::ParseFromParameter for $name {
+            fn parse_from_parameter(value: &str) -> poem_openapi::types::ParseResult<Self> {
                 Ok(Self(value.try_into()?))
             }
         }
 
-        impl ParseFromJSON for $name {
-            fn parse_from_json(value: Option<Value>) -> ParseResult<Self> {
+        #[cfg(feature = "poem")]
+        impl poem_openapi::types::ParseFromJSON for $name {
+            fn parse_from_json(value: Option<serde_json::Value>) -> poem_openapi::types::ParseResult<Self> {
                 match value {
-                    Some(Value::String(s)) => Ok(Self(Uuid::from_str(&s)?)),
+                    Some(serde_json::Value::String(s)) => Ok(Self(Uuid::from_str(&s)?)),
                     _ => Err(poem_openapi::types::ParseError::<$name>::custom(format!(
                         "Unexpected representation of {}",
                         stringify!($name)
@@ -131,9 +136,10 @@ macro_rules! newtype_uuid {
             }
         }
 
-        impl ToJSON for $name {
-            fn to_json(&self) -> Option<Value> {
-                Some(Value::String(self.0.to_string()))
+        #[cfg(feature = "poem")]
+        impl poem_openapi::types::ToJSON for $name {
+            fn to_json(&self) -> Option<serde_json::Value> {
+                Some(serde_json::Value::String(self.0.to_string()))
             }
         }
