+++ golem-common/src/retries.rs
@@ -18,7 +18,7 @@ use std::pin::Pin;
 use std::time::{Duration, Instant};
 use tracing::{error, info, warn, Level};
 
-use crate::config::RetryConfig;
+use crate::model::RetryConfig;
 use crate::metrics::external_calls::{
     record_external_call_failure, record_external_call_retry, record_external_call_success,
 };
@@ -221,7 +221,7 @@ where
 mod tests {
     use test_r::test;
 
-    use crate::config::RetryConfig;
+    use crate::model::RetryConfig;
     use std::time::Duration;
 
     #[test]
