+++ golem-rib/src/compiler/compiler_output.rs
@@ -30,8 +30,10 @@ pub struct CompilerOutput {
 
 #[cfg(feature = "protobuf")]
 mod protobuf {
+    use crate::{
+        CompilerOutput, RibByteCode, RibInputTypeInfo, RibOutputTypeInfo, WorkerFunctionsInRib,
+    };
     use golem_api_grpc::proto::golem::rib::CompilerOutput as ProtoCompilerOutput;
-    use crate::{CompilerOutput, RibByteCode, RibInputTypeInfo, RibOutputTypeInfo, WorkerFunctionsInRib};
 
     impl TryFrom<ProtoCompilerOutput> for CompilerOutput {
         type Error = String;
@@ -61,12 +63,14 @@ mod protobuf {
         }
     }
 
-    impl From<CompilerOutput> for ProtoCompilerOutput {
-        fn from(value: CompilerOutput) -> Self {
-            ProtoCompilerOutput {
-                byte_code: Some(golem_api_grpc::proto::golem::rib::RibByteCode::from(
+    impl TryFrom<CompilerOutput> for ProtoCompilerOutput {
+        type Error = String;
+
+        fn try_from(value: CompilerOutput) -> Result<Self, Self::Error> {
+            Ok(ProtoCompilerOutput {
+                byte_code: Some(golem_api_grpc::proto::golem::rib::RibByteCode::try_from(
                     value.byte_code,
-                )),
+                )?),
                 rib_input: Some(golem_api_grpc::proto::golem::rib::RibInputType::from(
                     value.rib_input_type_info,
                 )),
@@ -77,7 +81,7 @@ mod protobuf {
                 rib_output: value
                     .rib_output_type_info
                     .map(golem_api_grpc::proto::golem::rib::RibOutputType::from),
-            }
+            })
         }
     }
 }
