+++ golem-rib/src/interpreter/interpreter_stack_value.rs
@@ -157,8 +157,8 @@ mod internal {
     pub fn unable_to_complete_math_operation(left: &ValueAndType, right: &ValueAndType) -> String {
         format!(
             "Unable to complete math operation for {}, {}",
-            serde_json::to_string(left),
-            serde_json::to_string(right)
+            serde_json::to_string(left).unwrap_or_default(),
+            serde_json::to_string(right).unwrap_or_default()
         )
     }
